\name{glio}
\alias{glio}
\docType{data}
\title{ Sequencing data for a tumor sample with control. }
\description{
 A dataset of a paired tumor-normal sequencing sample from Complete Genomics.
}
\usage{glio()}
\format{
  A data frame with 3458745 rows on the following 7 variables.
  \describe{
    \item{\code{seqnames}}{a character vector}
    \item{\code{start}}{an integer vector}
    \item{\code{zygosity}}{a character vector}
    \item{\code{tCount}}{an integer vector}
    \item{\code{LAF}}{a numeric vector}
    \item{\code{tCountN}}{an integer vector}
    \item{\code{germLAF}}{a numeric vector}
  }
}

\keyword{datasets}
