\name{SomatiCAData-package}
\alias{SomatiCAData-package}
\alias{SomatiCAData}
\docType{package}
\title{
An example cancer whole genome sequencing data for the SomatiCA package
}
\description{
 A dataset of a paired tumor-normal sequencing sample from Complete Genomics.
}

\author{
Mengjie Chen <mengjie.chen@yale.edu>

Maintainer: Mengjie Chen <mengjie.chen@yale.edu>
}

\keyword{ package }

\examples{
data(glio)
}
